@@ -26,16 +26,37 @@
     "yargs": "^15.4.0"
   },
   "devDependencies": {
-    "eslint": "^7.4.0",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
+    "@types/node": "^16.11.6",
+    "eslint": "^8.16.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsdoc": "^39.3.1",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
+    "eslint-plugin-unicorn": "^42.0.0",
     "markdownlint": "^0.20.2",
     "markdownlint-cli": "^0.22.0",
     "mocha": "^10.0.0",
+    "gulp": "^4.0.2",
+    "gulp-eslint-new": "^1.4.4",
+    "gulp-typescript": "^6.0.0-alpha.1",
     "nyc": "^15.0.1",
+    "prettier": "2.6.2",
+    "prettier-eslint": "^15.0.0",
+    "typescript": "^4.7.2",
     "sinon": "^9.0.2"
   },
   "scripts": {
+    "gulp": "gulp",
     "test": "nyc mocha",
     "lint": "eslint \"./**/*.js\"",
+    "lint:fix": "eslint --fix src/ --ext=.ts,.tsx",
+    "prettier": "prettier --write \"src/**/*.{ts,tsx}\"",
+    "prettier:check": "prettier --check \"src/**/*.{ts,tsx}\"",
     "markdown": "markdownlint **/*.md --ignore node_modules"
   },
   "repository": {
