@@ -1,23 +1,24 @@
-// vim: tw=120 softtabstop=4 shiftwidth=4
-
-// see: https://github.com/online-go/gtp2ogs/blob/devel/docs/DEV.md#no-console-eslint-rule-in-configjs
-/* eslint-disable no-console */
-
-const { getArgNamesGRU } = require('./options/getArgNamesGRU');
-const { getOptionName } = require('./options/getOptionName');
-const { getRankedUnranked } = require('./options/getRankedUnranked');
-const { getRankedUnrankedUnderscored } = require('./options/getRankedUnrankedUnderscored');
-
-const { droppedOptions, ogsPvAIs, rankedUnrankedOptions } = require('./constants');
+import {
+    getArgNamesGRU,
+    getOptionName,
+    getRankedUnranked,
+    getRankedUnrankedUnderscored,
+} from "./options";
+import { droppedOptions, ogsPvAIs, rankedUnrankedOptions } from "./constants";
+import { trace } from "./trace";
+import { config } from "./constants";
+export { config } from "./constants";
 
 exportInitialConfig();
 
-exports.updateFromArgv = function(argv) {
+export function updateFromArgv(argv) {
     // console messages
     // A- greeting and debug status
 
     const debugStatus = argv.debug ? "DEBUG: ON\n" : "";
-    console.log(`\ngtp2ogs version 6.0.1\n--------------------\n- For changelog or latest devel updates, please visit https://github.com/online-go/gtp2ogs/tree/devel\n${debugStatus}`);
+    trace.log(
+        `\ngtp2ogs version 6.0.1\n--------------------\n- For changelog or latest devel updates, please visit https://github.com/online-go/gtp2ogs/tree/devel\n${debugStatus}`,
+    );
 
     // B - test unsupported argv
 
@@ -30,7 +31,7 @@ exports.updateFromArgv = function(argv) {
     testRankedUnrankedOptions(rankedUnrankedOptions, argv);
 
     // C - set general/ranked/unranked options defaults
-    
+
     // For general/ranked/unranked options, do not add a default using .default of yargs, add it later in the
     // code if no ranked arg nor unranked arg are used, else we would be force using the general arg regardless of
     // botadmin using the ranked and/or unranked arg(s), triggering the no 3 args at the same time error.
@@ -41,13 +42,13 @@ exports.updateFromArgv = function(argv) {
     // 0) Export everything in argv first
 
     for (const k in argv) {
-        exports[k] = argv[k];
+        config[k] = argv[k];
     }
 
     // 1) Add and Modify/Adjust exports
 
     if (argv.debug) {
-        exports.DEBUG = true;
+        config.DEBUG = true;
     }
     if (argv.logfile !== undefined) {
         exportLogfileFilename(argv.logfile, argv.debug);
@@ -56,19 +57,19 @@ exports.updateFromArgv = function(argv) {
     exportSecondsAsMilliseconds("startupbuffer", argv);
 
     if (argv.beta) {
-        exports.host = 'beta.online-go.com';
+        config.host = "beta.online-go.com";
     }
 
     exportNoAutoHandicapsIfMinHandicapsArePositive(argv);
 
     if (argv.ogspv) {
-        exports.ogspv = argv.ogspv.toUpperCase();
+        config.ogspv = argv.ogspv.toUpperCase();
     }
     if (argv.rejectnew) {
         // A bot never accepting new games shouldn't be appearing in the dropdown, polite to our users.
-        exports.hidden = true;
+        config.hidden = true;
     }
-    exports.bot_command = argv._;
+    config.bot_command = argv._;
 
     // 2) specific ranked/unranked options exports
 
@@ -81,34 +82,35 @@ exports.updateFromArgv = function(argv) {
 
     warnIfNopauseIsMissing("ranked");
     warnIfNopauseIsMissing("unranked");
-    console.log("\n");
-
+    trace.log("\n");
 }
+
+config.updateFromArgv = updateFromArgv;
 
 function exportInitialConfig() {
     exportInitialConfigUnderscored("");
     exportInitialConfigUnderscored("_ranked");
     exportInitialConfigUnderscored("_unranked");
 
-    exports.start_date = new Date();
+    config.start_date = new Date();
 }
 
 function exportInitialConfigUnderscored(rankedUnrankedUnderscored) {
-    exports[`banned_users${rankedUnrankedUnderscored}`] = {};
-    exports[`allow_all_boardsizes${rankedUnrankedUnderscored}`] = false;
-    exports[`allowed_boardsizes${rankedUnrankedUnderscored}`] = [];
-    exports[`allow_all_komis${rankedUnrankedUnderscored}`] = false;
-    exports[`allowed_komis${rankedUnrankedUnderscored}`] = [];
-    exports[`allowed_speeds${rankedUnrankedUnderscored}`] = {};
-    exports[`allow_all_speeds${rankedUnrankedUnderscored}`] = false;
-    exports[`allowed_timecontrols${rankedUnrankedUnderscored}`] = {};
-    exports[`allow_all_timecontrols${rankedUnrankedUnderscored}`] = false;
+    config[`banned_users${rankedUnrankedUnderscored}`] = {};
+    config[`allow_all_boardsizes${rankedUnrankedUnderscored}`] = false;
+    config[`allowed_boardsizes${rankedUnrankedUnderscored}`] = [];
+    config[`allow_all_komis${rankedUnrankedUnderscored}`] = false;
+    config[`allowed_komis${rankedUnrankedUnderscored}`] = [];
+    config[`allowed_speeds${rankedUnrankedUnderscored}`] = {};
+    config[`allow_all_speeds${rankedUnrankedUnderscored}`] = false;
+    config[`allowed_timecontrols${rankedUnrankedUnderscored}`] = {};
+    config[`allow_all_timecontrols${rankedUnrankedUnderscored}`] = false;
 }
 
 function testRankedUnrankedOptions(rankedUnrankedOptions, argv) {
     for (const option of rankedUnrankedOptions) {
         const [general, ranked, unranked] = getArgNamesGRU(option.name);
-        
+
         // check undefined specifically to handle valid values such as 0 or null which are tested false
         if (argv[general] !== undefined) {
             if (argv[ranked] !== undefined) {
@@ -140,14 +142,18 @@ function testDroppedArgv(droppedOptions, argv) {
     for (const [oldNames, newName] of droppedOptions) {
         for (const oldName of oldNames) {
             if (argv[oldName]) {
-                if (newName !== undefined) throw `Dropped: --${oldName} is no longer supported, use --${newName} instead.`;
-                throw `Dropped: --${oldName} is no longer supported.`;
+                if (newName !== undefined) {
+                    throw new Error(
+                        `Dropped: --${oldName} is no longer supported, use --${newName} instead.`,
+                    );
+                }
+                throw new Error(`Dropped: --${oldName} is no longer supported.`);
             }
         }
     }
     for (const argName of getArgNamesGRU("komis")) {
         if (argv[argName]) {
-            for (const komi of ["auto","null"]) {
+            for (const komi of ["auto", "null"]) {
                 if (argv[argName].split(",").includes(komi)) {
                     throw `Dropped: --${argName} ${komi} is no longer supported, use --${argName} automatic instead`;
                 }
@@ -164,28 +170,31 @@ function testConflictingOptions(optionNameFirst, optionNameSecond, argv) {
 
 function ensureSupportedOgspvAI(ogspv, ogsPvAIs) {
     // being case tolerant
-    if (!ogspv) return;
+    if (!ogspv) {
+        return;
+    }
     const upperCaseOgsPv = ogspv.toUpperCase();
-    const upperCaseAIs   = ogsPvAIs.map(e => e.toUpperCase());
+    const upperCaseAIs = ogsPvAIs.map((e) => e.toUpperCase());
 
     if (!upperCaseAIs.includes(upperCaseOgsPv)) {
-        throw `Unsupported --ogspv option ${ogspv}.\nSupported options are ${ogsPvAIs.join(', ')}`;
+        throw `Unsupported --ogspv option ${ogspv}.\nSupported options are ${ogsPvAIs.join(", ")}`;
     }
 }
 
 function setRankedUnrankedOptionsDefaults(rankedUnrankedOptions, argv) {
     for (const option of rankedUnrankedOptions) {
-        if (!("default" in option)) continue;
-        
+        if (!("default" in option)) {
+            continue;
+        }
+
         const [general, ranked, unranked] = getArgNamesGRU(option.name);
 
         if (argv[general] === undefined) {
-            if ((argv[ranked] === undefined) && (argv[unranked] === undefined)) {                
+            if (argv[ranked] === undefined && argv[unranked] === undefined) {
                 argv[general] = option.default;
             } else if (argv[ranked] !== undefined && argv[unranked] === undefined) {
                 // also set the opposite arg currently undefined to the default, more convenient for botadmin
                 argv[unranked] = option.default;
-
             } else if (argv[ranked] === undefined && argv[unranked] !== undefined) {
                 // also set the opposite arg currently undefined to the default, more convenient for botadmin
                 argv[ranked] = option.default;
@@ -195,7 +204,7 @@ function setRankedUnrankedOptionsDefaults(rankedUnrankedOptions, argv) {
 }
 
 function getLogfileFilename(argvLogfile) {
-    return (argvLogfile === "" ? `gtp2ogs-logfile-${exports.start_date.toISOString()}` : argvLogfile);
+    return argvLogfile === "" ? `gtp2ogs-logfile-${config.start_date.toISOString()}` : argvLogfile;
 }
 
 function getValidFilename(filename) {
@@ -208,17 +217,17 @@ function exportSecondsAsMilliseconds(optionName, argv) {
     if (argv[optionName] !== undefined) {
         // Convert some times in seconds to milliseconds once here so we don't need
         // to do it each time it is used later.
-        exports[optionName] = argv[optionName] * 1000;
+        config[optionName] = argv[optionName] * 1000;
     }
 }
 
-function exportNoAutoHandicapIfMinHandicapIsPositive(rankedUnranked, argv){
+function exportNoAutoHandicapIfMinHandicapIsPositive(rankedUnranked, argv) {
     // Setting minimum handicap higher than -1 has the consequence of rejecting
     // challenge that have automatic handicap in connection.js (-1 lower than
     // minimum handicap 0 or higher, rejecting challenge)
 
     if (argv[`minhandicap${rankedUnranked}`] > -1) {
-        exports[`noautohandicap${rankedUnranked}`] = true;
+        config[`noautohandicap${rankedUnranked}`] = true;
     }
 }
 
@@ -231,12 +240,14 @@ function exportNoAutoHandicapsIfMinHandicapsArePositive(argv) {
 function exportLogfileFilename(argvLogfile, argvDebug) {
     const filename = getLogfileFilename(argvLogfile);
     const validFilename = getValidFilename(filename);
-    
+
     if (argvDebug && filename !== validFilename) {
-        console.log (`Logfile name "${filename}" has been automatically renamed to "${validFilename}".\nValid logfile name can only be composed of letters (A-Z a-z), numbers (0-9), hyphens (-), underscores (_), spaces ( ), dots (.).\n`);
+        trace.log(
+            `Logfile name "${filename}" has been automatically renamed to "${validFilename}".\nValid logfile name can only be composed of letters (A-Z a-z), numbers (0-9), hyphens (-), underscores (_), spaces ( ), dots (.).\n`,
+        );
     }
 
-    exports.logfile = validFilename;
+    config.logfile = validFilename;
 }
 
 function parseRank(arg) {
@@ -246,11 +257,11 @@ function parseRank(arg) {
 
         if (results) {
             if (results[2] === "k") {
-                return (30 - parseInt(results[1]));
+                return 30 - parseInt(results[1]);
             } else if (results[2] === "d") {
-                return (30 - 1 + parseInt(results[1]));
+                return 30 - 1 + parseInt(results[1]);
             } else if (results[2] === "p") {
-                return (36 + parseInt(results[1]));
+                return 36 + parseInt(results[1]);
             }
         } else {
             throw `error: could not parse rank -${arg}-`;
@@ -261,7 +272,7 @@ function parseRank(arg) {
 function processRankExport(argName, argv) {
     const arg = argv[argName];
     if (arg) {
-        exports[argName] = parseRank(arg);
+        config[argName] = parseRank(arg);
     }
 }
 
@@ -280,9 +291,9 @@ function processBansExport(argName, argv) {
     const rankedUnrankedUnderscored = getRankedUnrankedUnderscored(argName);
 
     if (arg) {
-        const bans = arg.split(',');
+        const bans = arg.split(",");
         for (const bannedUser of bans) {
-            exports[`banned_users${rankedUnrankedUnderscored}`][bannedUser] = true;
+            config[`banned_users${rankedUnrankedUnderscored}`][bannedUser] = true;
         }
     }
 }
@@ -299,12 +310,12 @@ function processBoardsizesExport(argName, argv) {
     if (arg) {
         const rankedUnranked = getRankedUnranked(argName);
         const rankedUnrankedUnderscored = getRankedUnrankedUnderscored(rankedUnranked);
-        const boardsizes = arg.split(',');
+        const boardsizes = arg.split(",");
         for (const boardsize of boardsizes) {
             if (boardsize === "all") {
-                exports[`allow_all_boardsizes${rankedUnrankedUnderscored}`] = true;
+                config[`allow_all_boardsizes${rankedUnrankedUnderscored}`] = true;
             } else {
-                exports[`allowed_boardsizes${rankedUnrankedUnderscored}`][boardsize] = true;
+                config[`allowed_boardsizes${rankedUnrankedUnderscored}`][boardsize] = true;
             }
         }
     }
@@ -316,17 +327,17 @@ function processKomisExport(argName, argv) {
     if (arg) {
         const rankedUnranked = getRankedUnranked(argName);
         const rankedUnrankedUnderscored = getRankedUnrankedUnderscored(rankedUnranked);
-        const komis = arg.split(',');
+        const komis = arg.split(",");
         for (const komi of komis) {
             if (komi === "all") {
-                exports[`allow_all_komis${rankedUnrankedUnderscored}`] = true;
+                config[`allow_all_komis${rankedUnrankedUnderscored}`] = true;
             } else if (komi === "automatic") {
-                exports[`allowed_komis${rankedUnrankedUnderscored}`][null] = true;
+                config[`allowed_komis${rankedUnrankedUnderscored}`]["null"] = true;
             } else {
-                exports[`allowed_komis${rankedUnrankedUnderscored}`][komi] = true;
+                config[`allowed_komis${rankedUnrankedUnderscored}`][komi] = true;
             }
         }
-    } 
+    }
 }
 
 function processAllowedGroupExport(argName, argv) {
@@ -336,15 +347,15 @@ function processAllowedGroupExport(argName, argv) {
         const optionName = getOptionName(argName);
         const rankedUnranked = getRankedUnranked(argName);
         const rankedUnrankedUnderscored = getRankedUnrankedUnderscored(rankedUnranked);
-        const allowedValues = arg.split(',');
+        const allowedValues = arg.split(",");
         for (const allowedValue of allowedValues) {
             if (allowedValue === "all") {
-                exports[`allow_all_${optionName}${rankedUnrankedUnderscored}`] = true;
+                config[`allow_all_${optionName}${rankedUnrankedUnderscored}`] = true;
             } else {
-                exports[`allowed_${optionName}${rankedUnrankedUnderscored}`][allowedValue] = true;
+                config[`allowed_${optionName}${rankedUnrankedUnderscored}`][allowedValue] = true;
             }
         }
-    } 
+    }
 }
 
 function exportAllowedGroups(argv) {
@@ -367,7 +378,9 @@ function exportAllowedGroups(argv) {
 
 function warnIfNopauseIsMissing(rankedUnranked) {
     // avoid infinite games
-    if (!exports.nopause && !exports[`nopause${rankedUnranked}`]) {
-        console.log(`Warning: No --nopause nor --nopause${rankedUnranked}, ${rankedUnranked} games are likely to last forever`); 
+    if (!config.nopause && !config[`nopause${rankedUnranked}`]) {
+        trace.log(
+            `Warning: No --nopause nor --nopause${rankedUnranked}, ${rankedUnranked} games are likely to last forever`,
+        );
     }
 }
