@@ -1,27 +1,24 @@
 #!/usr/bin/env node
-// vim: tw=120 softtabstop=4 shiftwidth=4
-
-'use strict';
 
 // Remove api key from command line ASAP.
-process.title = 'gtp2ogs';
+process.title = "gtp2ogs";
 
 // Do this before importing anything else in case the other modules use config.
-const { getArgv } = require('./getArgv');
+import { getArgv } from "./getArgv";
+import { config } from "./config";
+
 const argv = getArgv();
-const config = require('./config');
 config.updateFromArgv(argv);
 
-process.title = `gtp2ogs ${config.bot_command.join(' ')}`;
+process.title = `gtp2ogs ${config.bot_command.join(" ")}`;
 
-const { console } = require('./console');
+import { trace } from "trace";
+import { io } from "socket.io-client";
+import { Connection } from "./Connection";
 
-const io = require('socket.io-client');
-const { Connection } = require('./connection');
-
-process.on('uncaughtException', function (er) {
+process.on("uncaughtException", (er) => {
     console.trace("ERROR: Uncaught exception");
-    console.error(`ERROR: ${er.stack}`);
+    trace.error(`ERROR: ${er.stack}`);
     if (!conn || !conn.socket) {
         conn = getNewConnection();
     } else {
